Fix surface size reporting
