+++ lib/src/cursor.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/services.dart';
+
+const Map<String, SystemMouseCursor> _cursors = {
+  'none': SystemMouseCursors.none,
+  'basic': SystemMouseCursors.basic,
+  'click': SystemMouseCursors.click,
+  'forbidden': SystemMouseCursors.forbidden,
+  'wait': SystemMouseCursors.wait,
+  'progress': SystemMouseCursors.progress,
+  'contextMenu': SystemMouseCursors.contextMenu,
+  'help': SystemMouseCursors.help,
+  'text': SystemMouseCursors.text,
+  'verticalText': SystemMouseCursors.verticalText,
+  'cell': SystemMouseCursors.cell,
+  'precise': SystemMouseCursors.precise,
+  'move': SystemMouseCursors.move,
+  'grab': SystemMouseCursors.grab,
+  'grabbing': SystemMouseCursors.grabbing,
+  'noDrop': SystemMouseCursors.noDrop,
+  'alias': SystemMouseCursors.alias,
+  'copy': SystemMouseCursors.copy,
+  'disappearing': SystemMouseCursors.disappearing,
+  'allScroll': SystemMouseCursors.allScroll,
+  'resizeLeftRight': SystemMouseCursors.resizeLeftRight,
+  'resizeUpDown': SystemMouseCursors.resizeUpDown,
+  'resizeUpLeftDownRight': SystemMouseCursors.resizeUpLeftDownRight,
+  'resizeUpRightDownLeft': SystemMouseCursors.resizeUpRightDownLeft,
+  'resizeUp': SystemMouseCursors.resizeUp,
+  'resizeDown': SystemMouseCursors.resizeDown,
+  'resizeLeft': SystemMouseCursors.resizeLeft,
+  'resizeRight': SystemMouseCursors.resizeRight,
+  'resizeUpLeft': SystemMouseCursors.resizeUpLeft,
+  'resizeUpRight': SystemMouseCursors.resizeUpRight,
+  'resizeDownLeft': SystemMouseCursors.resizeDownLeft,
+  'resizeDownRight': SystemMouseCursors.resizeDownRight,
+  'resizeColumn': SystemMouseCursors.resizeColumn,
+  'resizeRow': SystemMouseCursors.resizeRow,
+  'zoomIn': SystemMouseCursors.zoomIn,
+  'zoomOut': SystemMouseCursors.zoomOut,
+};
+
+SystemMouseCursor getCursorByName(String name) =>
+    _cursors[name] ?? SystemMouseCursors.basic;
